requirements.txt
requests
pandas
openpyxl
requests
psycopg2-binary
tqdm

